Moved mochi.py to builtins.py.
