+++ mochi/core/__init__.py
@@ -1,1 +1,1 @@
-from .mochi import *
+from .builtins import *
